api/requirements.txt
affine==2.4.0
shapely==2.0.1
geotiff==0.2.9
geopandas==0.13.0
scipy==1.10.1
pyproj==3.5.0
rasterio==1.3.6
pykml==0.2.0
pandas==1.3.5
matplotlib==3.7.1
matplotlib-inline==0.1.3
numpy==1.21.5
folium==0.14.0
sentinelsat==1.2.1
selenium==4.1.0
selenium-wire
Flask==2.2.2
Flask-Session==0.3.2
passlib==1.7.2
peewee==3.14.1
peewee-migrate==1.1.6
psycopg2-binary==2.8.6
pytz==2021.1
redis==3.5.3
requests==2.27.1
mechanize==0.4.7
Pillow==9.0.1
pytesseract==0.3.8
requests_html==0.10.0
urllib3==1.26.0
itsdangerous==2.0.1
shapely==2.0.1
SQLAlchemy==1.4.46
